refactor: pages내부에 savingCalculator 디렉토리 추가
